backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
passlib[bcrypt]
python-jose
pytest
pytest-asyncio
httpx
openai
transformers
